+++ crypto/bio/bss_log.c
@@ -240,7 +240,7 @@ static void xsyslog(BIO *bp, int priority, const char *string)
         break;
     }
 
-    BIO_snprintf(pidbuf, sizeof(pidbuf), "[%lu] ", GetCurrentProcessId());
+    snprintf(pidbuf, sizeof(pidbuf), "[%lu] ", GetCurrentProcessId());
     lpszStrings[0] = pidbuf;
     lpszStrings[1] = string;