+++ crypto/x509/v3_san.c
@@ -128,7 +128,7 @@ STACK_OF(CONF_VALUE) *i2v_GENERAL_NAME(X509V3_EXT_METHOD *method,
             break;
         default:
             if (OBJ_obj2txt(oline, sizeof(oline), gen->d.otherName->type_id, 0) > 0)
-                BIO_snprintf(othername, sizeof(othername), "othername: %s",
+                snprintf(othername, sizeof(othername), "othername: %s",
                     oline);
             else
                 OPENSSL_strlcpy(othername, "othername", sizeof(othername));